add tests to thumbnails app
