+++ mapstory/test.py
@@ -20,4 +20,4 @@ from tests.testMapstory import *
 from journal.tests import *
 from apps.boxes.tests import *
 from apps.flag.tests import *
-
+from apps.thumbnails.tests import *
